feat(PhasePlanner): Integrate bounties into phase planner view
